[utils] Merge all utils code into index.ts
## Summary

Speed up type checking a little bit by avoiding one extra lookup

## Test Plan

CI
